example/requirements.txt
pyzmq==22.3.0
msgpack==1.0.2
opencv-python==4.5.3.56